@@ -1,94 +1,101 @@
 name: litelearninglab
 description: A new Flutter application.
-
 publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  firebase_core: ^2.14.0
-  flutter_card_swiper: ^7.0.1
-  smooth_page_indicator: ^1.1.0
-  showcaseview: ^3.0.0
-  expansion_tile_group:
-  cupertino_icons: ^1.0.5
-  after_layout: ^1.2.0
-  cloud_firestore: ^4.8.1
-  firebase_auth: ^4.6.3
-  intl: 0.17.0
-  toast: ^0.3.0
-  provider: ^6.0.5
-  url_launcher: ^6.1.11
-  firebase_database: ^10.2.3
-  firebase_messaging: ^14.6.3
-  audioplayers: ^4.1.0
-  speech_to_text: ^6.1.1
-  number_to_words: ^1.0.0
-  sqflite: ^2.2.8+4 #no web
-  shared_preferences: ^2.1.2
-  flutter_tts: ^3.7.0
-  animations: ^2.0.7
-  flutter_inappwebview: ^5.7.2+3 #no web
-  device_info: #no web
-  package_info: ^2.0.2 #no web
-  get: ^4.6.5
-  permission_handler:
-  cached_network_image: ^3.2.3
-  pin_code_fields: ^7.4.0
-#  flutter_statusbarcolor: ^0.2.3
-  scroll_to_index: ^3.0.1
-  aes_crypt_null_safe: ^2.0.2
-  path_provider: ^2.0.15
-  firebase_remote_config: ^4.2.3
-  video_player: ^2.1.6
-  chewie: ^1.8.1
-  flutter_local_notifications: ^14.1.1
-  firebase_analytics: ^10.4.3
-  flutter_svg: ^2.0.9
-  number_to_words_english: ^2.0.2
-  auto_size_text: ^3.0.0
+  # Firebase
+  firebase_core: ^3.9.0
+  cloud_firestore: ^5.6.0
+  firebase_auth: ^5.3.1
+  firebase_database: ^11.3.1
+  firebase_messaging: ^15.1.3
+  firebase_remote_config: ^5.1.4
+  firebase_analytics: ^11.3.2
+  firebase_crashlytics: ^4.1.3
+  firebase_app_installations: ^0.3.1+6
 
-  percent_indicator: ^4.2.3
-  google_fonts: ^4.0.4
+  # Core packages
+  provider: ^6.1.2
+  get: ^4.6.6
+  intl: ^0.20.2
+  http: ^1.2.2
+  path_provider: ^2.1.5
+  shared_preferences: ^2.3.2
+
+  # UI / UX
+  flutter_card_swiper: ^7.0.1
+  smooth_page_indicator: ^1.2.0
+  showcaseview: ^4.0.1
+  expansion_tile_group: ^2.3.0
+  cupertino_icons: ^1.0.8
+  audioplayers: ^6.0.0
+  # speech_to_text: ^7.3.0
+  # flutter_tts: ^4.0.2
+  animations: ^2.0.11
+  cached_network_image: ^3.4.1
+  pin_code_fields: ^8.0.1
+  scroll_to_index: ^3.0.1
+  video_player: ^2.9.2
+  chewie: ^1.8.5
+  flutter_local_notifications: ^18.0.1
+  flutter_svg: ^2.0.10+1
+  auto_size_text: ^3.0.0
+  percent_indicator: ^4.2.4
+  google_fonts: ^6.2.1
   card_swiper: ^3.0.1
-  upgrader:
+  upgrader: ^11.0.1
   flutter_page_view_indicator: ^0.0.5
   card_stack_widget: ^0.1.5
-  swipe_deck: ^1.0.0
+  swipe_deck: ^1.1.0
   slide_switcher: ^1.1.2
-  flutter_colorful_tab: ^0.1.1
-  http: ^0.13.6
-  json_annotation: ^4.9.0
   hive: ^2.2.3
   hive_flutter: ^1.1.0
-  lottie:
-  fl_chart: ^0.69.0
-  syncfusion_flutter_charts: 20.3.61
-  connectivity_plus: ^6.0.3
+  lottie: ^3.2.0
+  fl_chart: ^1.1.0
+  syncfusion_flutter_charts: ^30.2.7
+  connectivity_plus: ^6.1.0
   flutter_user_guildance: ^2.0.7
+  flutter_cache_manager: ^3.4.1
+  flick_video_player: ^0.9.0
+  encrypt: ^5.0.3
+  device_info_plus: ^11.3.0
+  package_info_plus: ^8.0.2
+  flutter_inappwebview: ^6.0.0-beta.30
+  after_layout: ^1.2.0
+  number_to_words_english: ^3.0.4
+  permission_handler: ^12.0.1
+  number_to_words: ^1.0.0
+  aes_crypt_null_safe: ^3.1.1
+  flutter_timezone: ^4.1.1
+  google_nav_bar: ^5.0.7
+  toast: ^0.3.0
+  flutter_tts: ^4.2.3
+  speech_to_text: ^7.3.0
+  url_launcher: ^6.3.2
+  android_id: ^0.4.0
+  # flutter_colorful_tab: ^0.1.1
+  # flutter_colorful_tab: ^0.1.1
 
 
-
-
-
-
-
+dependency_overrides:
+  shared_preferences: ^2.3.2
+  shared_preferences_android: ^2.3.2
 
 dev_dependencies:
-  build_runner: ^2.1.0
+  build_runner: ^2.4.9
   hive_generator: ^2.0.1
-  json_serializable: ^6.6.1
+  json_serializable: ^6.7.1
   flutter_test:
     sdk: flutter
 
-
 flutter:
-
   uses-material-design: true
 
   assets:
