@@ -6,14 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 // import 'package:flutter_statusbarcolor/flutter_statusbarcolor.dart';
 import 'package:get/get_navigation/src/root/get_material_app.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:hive/hive.dart';
-import 'package:hive_flutter/adapters.dart';
-import 'package:litelearninglab/API/api_functions.dart';
-import 'package:litelearninglab/screens/dashboard/dashboard_screen.dart';
-import 'package:litelearninglab/screens/dashboard/new_dashboard.dart';
 import 'package:litelearninglab/screens/dashboard/school_dashboard.dart';
-import 'package:litelearninglab/screens/login/login_screen.dart';
 import 'package:litelearninglab/screens/login/new_login_screen.dart';
 import 'package:litelearninglab/screens/login/unauth_screen.dart';
 import 'package:litelearninglab/screens/process_learning/indicator_controller.dart';
@@ -24,15 +17,12 @@ import 'package:litelearninglab/states/auth_state.dart';
 import 'package:litelearninglab/utils/bottom_navigation.dart';
 import 'package:litelearninglab/utils/commonfunctions/common_functions.dart';
 import 'package:litelearninglab/utils/fcm.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:provider/provider.dart';
 import 'package:toast/toast.dart';
+
 import 'config/app_config.dart';
 import 'constants/enums.dart';
 import 'constants/keys.dart';
-import 'constants/strings.dart';
-import 'hiveDb/hiveDb.dart';
-import 'hiveDb/new_process_hive_adapter.dart';
 
 const Map<int, Color> colorSwatch = {
   50: Color(0xFF293750),
@@ -54,8 +44,11 @@ Future<void> main() async {
   print("hejeo e hh uh ");
   await Firebase.initializeApp();
 
-  SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(systemNavigationBarColor: Color(0xFF293750), statusBarColor: Color(0xFF293750)));
-  var configuredApp = new AppConfig(appName: 'Profluent', flavorName: 'prod', fcmKey: '', child: MyApp());
+  SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
+      systemNavigationBarColor: Color(0xFF293750),
+      statusBarColor: Color(0xFF293750)));
+  var configuredApp = new AppConfig(
+      appName: 'Profluent', flavorName: 'prod', fcmKey: '', child: MyApp());
   runApp(
     MultiProvider(
       providers: [
@@ -76,6 +69,7 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
   @override
   void initState() {
     super.initState();
+
     WidgetsBinding.instance.addObserver(this); // Add lifecycle observer
   }
 
@@ -87,28 +81,24 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
-    if (state == AppLifecycleState.paused) {
-      subResume = false;
-      resume = false;
-      log("User navigated to another app or attended a call.");
+    if (state == AppLifecycleState.paused ||
+        state == AppLifecycleState.inactive ||
+        state == AppLifecycleState.detached) {
+      if (count > 0) {
+        recordTiming("Paused");
+      }
+      log("User navigated to another app or attended a call.${count}");
     } else if (state == AppLifecycleState.resumed) {
+      startTimings = DateTime.now();
       resume = true;
       subResume = true;
       log("User returned to the app.");
-    } else if (state == AppLifecycleState.inactive) {
-      resume = false;
-      subResume = false;
-      log("App is inactive (e.g., during a phone call).");
-    } else if (state == AppLifecycleState.detached) {
-      resume = false;
-      subResume = false;
-      log("App is closed or detached.");
     }
   }
 
   Widget build(BuildContext context) {
     Fcm.getInstance(context)?.initConfigure();
-
+    // checkUserAndCompanyStatus(context);
     //  SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual, overlays: [SystemUiOverlay.top]);
     // SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersive);
     // SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(statusBarColor: Color(0xFF293750)));
@@ -117,6 +107,7 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
       debugShowCheckedModeBanner: false,
       title: 'Profluent',
       theme: ThemeData(
+          scaffoldBackgroundColor: Color(0xFF293750),
           fontFamily: Keys.fontFamily,
           splashColor: Colors.transparent,
           highlightColor: Colors.transparent,
@@ -127,82 +118,84 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
             },
           ),
           primaryColor: Color(0xFF293750),
-          primaryIconTheme: Theme.of(context).primaryIconTheme.copyWith(color: Colors.black),
+          primaryIconTheme:
+              Theme.of(context).primaryIconTheme.copyWith(color: Colors.black),
           primaryTextTheme: Theme.of(context).primaryTextTheme.apply(
                 bodyColor: Colors.black,
                 fontFamily: Keys.fontFamily,
               ),
           textTheme: TextTheme(
-            bodyText1: Theme.of(context).textTheme.bodyText1?.merge(
+            bodyLarge: Theme.of(context).textTheme.bodyLarge?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            bodyText2: Theme.of(context).textTheme.bodyText2?.merge(
+            bodyMedium: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            button: Theme.of(context).textTheme.bodyText2?.merge(
+            labelLarge: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            headline1: Theme.of(context).textTheme.bodyText2?.merge(
+            displayLarge: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            headline2: Theme.of(context).textTheme.bodyText2?.merge(
+            displayMedium: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            headline3: Theme.of(context).textTheme.bodyText2?.merge(
+            displaySmall: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            headline4: Theme.of(context).textTheme.bodyText2?.merge(
+            headlineMedium: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            headline5: Theme.of(context).textTheme.bodyText2?.merge(
+            headlineSmall: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            headline6: Theme.of(context).textTheme.bodyText2?.merge(
+            titleLarge: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            subtitle1: Theme.of(context).textTheme.bodyText2?.merge(
+            titleMedium: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            subtitle2: Theme.of(context).textTheme.bodyText2?.merge(
+            titleSmall: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            caption: Theme.of(context).textTheme.bodyText2?.merge(
+            bodySmall: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
-            overline: Theme.of(context).textTheme.bodyText2?.merge(
+            labelSmall: Theme.of(context).textTheme.bodyMedium?.merge(
                   TextStyle(
                     fontFamily: Keys.fontFamily,
                   ),
                 ),
           ),
-          colorScheme: ColorScheme.fromSwatch(primarySwatch: customColor).copyWith(secondary: customColor),
+          colorScheme: ColorScheme.fromSwatch(primarySwatch: customColor)
+              .copyWith(secondary: customColor),
           scrollbarTheme: ScrollbarThemeData().copyWith(
-            thumbColor: MaterialStateProperty.all(Colors.white),
-            thumbVisibility: MaterialStateProperty.all<bool>(true),
+            thumbColor: WidgetStateProperty.all(Colors.white),
+            thumbVisibility: WidgetStateProperty.all<bool>(true),
           )),
       home: AuthWrapper(),
       scrollBehavior: RemoveGlowEffect(),
@@ -221,13 +214,17 @@ class AuthWrapper extends StatelessWidget {
       builder: (context, snapshot) {
         if (snapshot.connectionState == ConnectionState.waiting) {
           // Show a loading screen if the authentication state is still unknown
-          return Scaffold(body: Center(child: CircularProgressIndicator(color: Colors.white)));
+          return Scaffold(
+              body: Center(
+                  child: CircularProgressIndicator(color: Colors.white)));
         } else {
           print("shdofofaoufaf : ${snapshot.data}");
 
           switch (snapshot.data) {
             case Status.authenticated:
-              return config?.flavorName == "english" ? SchoolDashboard() : BottomNavigation();
+              return config?.flavorName == "english"
+                  ? SchoolDashboard()
+                  : BottomNavigation();
             case Status.tutorial:
               return TutorialScreen(
                 url: 'assets/mp4_video/video_20241022_115109_check.mp4',
@@ -238,11 +235,13 @@ class AuthWrapper extends StatelessWidget {
               return NewLoginScreen();
             case Status.userNotExist:
               return UnauthScreen(
-                text: 'USER NOT EXISTS!, Please contact administrator for more details.',
+                text:
+                    'USER NOT EXISTS!, Please contact administrator for more details.',
               );
             case Status.userInactive:
               return UnauthScreen(
-                text: 'USER NOT ACTIVE!, Please contact administrator for more details',
+                text:
+                    'USER NOT ACTIVE!, Please contact administrator for more details',
               );
             case Status.deviceChanged:
               return UnauthScreen(
@@ -264,7 +263,8 @@ class AuthWrapper extends StatelessWidget {
 
 class RemoveGlowEffect extends ScrollBehavior {
   @override
-  Widget buildOverscrollIndicator(BuildContext context, Widget child, ScrollableDetails details) {
+  Widget buildOverscrollIndicator(
+      BuildContext context, Widget child, ScrollableDetails details) {
     return child;
   }
 }
