@@ -1,6 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-
-import '../utils/FirestoreService.dart';
+import 'package:litelearninglab/utils/FirestoreService.dart';
 
 class UserM implements Jsonable {
   String? id;
@@ -18,6 +17,9 @@ class UserM implements Jsonable {
   String? fcmID;
   String? access;
   String? fcmKey;
+  String? lastLogin;
+  String? companyId;
+  String? firstTImeLogin;
 
   UserM(
       {this.company,
@@ -34,9 +36,12 @@ class UserM implements Jsonable {
       this.city,
       this.encryptionKey,
       this.UserMname,
-      this.fcmKey});
+      this.fcmKey,
+      this.lastLogin,
+      this.companyId,
+      this.firstTImeLogin});
 
-  toJson() {
+  Map<String, dynamic> toJson() {
     return {
       "company": company,
       "email": email,
@@ -52,17 +57,21 @@ class UserM implements Jsonable {
       "username": UserMname,
       "access": access,
       "fcmKey": fcmKey,
+      "lastLogin": lastLogin,
+      "companyid": companyId,
+      'firstTimeLogin': firstTImeLogin
+      // Consistent lowercase
     };
   }
 
   UserM.map(dynamic obj) {
     this.id = obj['id'];
     this.company = obj['company'];
-
+    this.lastLogin = obj['lastLogin'];
+    this.companyId = obj['companyid']; // Consistent lowercase
     this.email = obj['email'];
     this.imei = obj['imei'];
     this.mobile = obj['mobile'];
-
     this.model = obj['model'];
     this.profile = obj['profile'];
     this.city = obj['city'];
@@ -71,104 +80,97 @@ class UserM implements Jsonable {
     this.UserMname = obj['username'];
     this.access = obj['access'];
     this.fcmKey = obj['fcmKey'];
+    this.firstTImeLogin = obj['firstTimeLogin'];
   }
 
   Map<String, dynamic> toMap() {
-    var map = new Map<String, dynamic>();
-    if (id != null) {
-      map['id'] = id;
-    }
-    if (company != null) {
-      map['company'] = company;
-    }
-    if (fcmID != null) {
-      map['fcmID'] = fcmID;
-    }
-    if (email != null) {
-      map['email'] = email;
-    }
-    if (access != null) {
-      map['access'] = access;
-    }
-    if (imei != null) {
-      map['imei'] = imei;
-    }
-    if (mobile != null) {
-      map['mobile'] = mobile;
-    }
-    if (model != null) {
-      map['model'] = model;
-    }
-    if (profile != null) {
-      map['profile'] = profile;
-    }
-    if (city != null) {
-      map['city'] = city;
-    }
-    if (team != null) {
-      map['team'] = team;
-    }
-    if (status != null) {
-      map['status'] = status;
-    }
-    if (UserMname != null) {
-      map['username'] = UserMname;
-    }
-    if (fcmKey != null) {
-      map['fcmKey'] = fcmKey;
-    }
+    var map = <String, dynamic>{};
+    if (id != null) map['id'] = id;
+    if (companyId != null) map['companyid'] = companyId; // Consistent lowercase
+    if (company != null) map['company'] = company;
+    if (lastLogin != null) map['lastLogin'] = lastLogin;
+    if (fcmID != null) map['fcmID'] = fcmID;
+    if (email != null) map['email'] = email;
+    if (access != null) map['access'] = access;
+    if (imei != null) map['imei'] = imei;
+    if (mobile != null) map['mobile'] = mobile;
+    if (model != null) map['model'] = model;
+    if (profile != null) map['profile'] = profile;
+    if (city != null) map['city'] = city;
+    if (team != null) map['team'] = team;
+    if (status != null) map['status'] = status;
+    if (UserMname != null) map['username'] = UserMname;
+    if (fcmKey != null) map['fcmKey'] = fcmKey;
+    if (firstTImeLogin != null) map['firstTimeLogin'];
     return map;
   }
 
   factory UserM.fromJson(DocumentSnapshot doc) {
-    Map map = doc.data() as Map;
+    final map = doc.data() as Map<String, dynamic>;
+    print('Firestore Document Fields: ${map.keys}'); // Debugging
+
     return UserM(
-      id: doc.id,
-      company: map['company'],
-      email: map['email'],
-      imei: map['imei'],
-      fcmID: map['fcmID'],
-      access: map['access'],
-      mobile: map['mobile'],
-      city: map['city'] is List ? (map['city'] as List).join(', ') : map['city'],
-      model: map['model'],
-      team: map['team'] is List ? (map['team'] as List).join(', ') : map['team'],
-      profile: map['profile'],
-      status: map['status'],
-      UserMname: map['username'],
-      fcmKey: map['fcmKey'],
-    );
+        id: doc.id,
+        company: map['company'],
+        lastLogin: map['lastLogin'],
+        email: map['email'],
+        companyId: map['companyid'] ?? '', // Consistent name with fallback
+        imei: map['imei'],
+        fcmID: map['fcmID'],
+        access: map['access'],
+        mobile: map['mobile'],
+        city: map['city'] is List
+            ? (map['city'] as List).join(', ')
+            : map['city'],
+        model: map['model'],
+        team: map['team'] is List
+            ? (map['team'] as List).join(', ')
+            : map['team'],
+        profile: map['profile'],
+        status: map['status'],
+        UserMname: map['username'],
+        fcmKey: map['fcmKey'],
+        firstTImeLogin: map['firstTimeLogin']);
   }
 
   UserM.fromMap(Map<String, dynamic> map)
-      : this.id = map['id'],
-        this.company = map['company'],
-        this.email = map['email'],
-        this.imei = map['imei'],
-        this.mobile = map['mobile'],
-        this.model = map['model'],
-        this.profile = map['profile'],
-        this.fcmID = map['fcmID'],
-        this.city = map['city'] is List ? (map['city'] as List).join(', ') : map['city'],
-        this.team = map['team'] is List ? (map['team'] as List).join(', ') : map['team'],
-        this.status = map['status'],
-        this.access = map['access'],
-        this.UserMname = map['username'],
-        this.fcmKey = map['fcmKey'];
+      : id = map['id'],
+        company = map['company'],
+        lastLogin = map['lastLogin'],
+        email = map['email'],
+        imei = map['imei'],
+        mobile = map['mobile'],
+        model = map['model'],
+        profile = map['profile'],
+        fcmID = map['fcmID'],
+        city = map['city'] is List
+            ? (map['city'] as List).join(', ')
+            : map['city'],
+        team = map['team'] is List
+            ? (map['team'] as List).join(', ')
+            : map['team'],
+        status = map['status'],
+        access = map['access'],
+        UserMname = map['username'],
+        fcmKey = map['fcmKey'],
+        companyId = map['companyid'] ?? '',
+        firstTImeLogin = map['firstTimeLogin'] ?? '';
+  // Consistent with fallback
 
-// UserM.fromMap(Map<String, dynamic> map) {
-//   this.key = map['key'];
-//   this.company = map['company'];
-//
-//   this.email = map['email'];
-//   this.imei = map['imei'];
-//
-//   this.mobile = map['mobile'];
-//   this.model = map['model'];
-//   this.profile = map['profile'];
-//   this.city = map['city'];
-//   this.team = map['team'];
-//   this.status = map['status'];
-//   this.UserMname = map['username'];
-// }
+  // Helper method to update active users count
+  static Future<void> incrementActiveUsers(String companyId) async {
+    if (companyId.isEmpty) return;
+
+    try {
+      await FirebaseFirestore.instance
+          .collection('userNode')
+          .doc(companyId)
+          .update({
+        'activeusers': FieldValue.increment(1),
+      });
+    } catch (e) {
+      print('Error incrementing active users: $e');
+      // Handle error or rethrow if needed
+    }
+  }
 }
