@@ -4,10 +4,8 @@ import 'package:get/get.dart';
 import 'package:litelearninglab/API/api.dart';
 import 'package:litelearninglab/common_widgets/spacings.dart';
 import 'package:litelearninglab/constants/app_colors.dart';
-import 'package:litelearninglab/main.dart';
 import 'package:litelearninglab/screens/webview/webview_screen.dart';
 import 'package:litelearninglab/states/auth_state.dart';
-import 'package:litelearninglab/utils/audio_player_manager.dart';
 import 'package:litelearninglab/utils/bottom_navigation.dart';
 import 'package:litelearninglab/utils/commonfunctions/common_functions.dart';
 import 'package:litelearninglab/utils/sizes_helpers.dart';
@@ -23,7 +21,8 @@ import '../../utils/shared_pref.dart';
 import '../word_screen/widgets/drop_down_word_item.dart';
 
 class InteracticeCatScreen extends StatefulWidget {
-  InteracticeCatScreen({Key? key, required this.linkCats, required this.title}) : super(key: key);
+  InteracticeCatScreen({Key? key, required this.linkCats, required this.title})
+      : super(key: key);
   final List<InteractiveLink> linkCats;
   final String title;
 
@@ -43,7 +42,10 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
   void initState() {
     startTimerSubCategory(arCallSimulation, widget.title);
     // isPlaying = List.generate(widget.linkCats.length, (index) => false);
-    controller = AutoScrollController(viewportBoundaryGetter: () => Rect.fromLTRB(0, 0, 0, MediaQuery.of(context).padding.bottom), axis: Axis.vertical);
+    controller = AutoScrollController(
+        viewportBoundaryGetter: () =>
+            Rect.fromLTRB(0, 0, 0, MediaQuery.of(context).padding.bottom),
+        axis: Axis.vertical);
     super.initState();
   }
 
@@ -70,7 +72,9 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                     ),
                   ),
                   Container(
-                    height: isSplitScreen ? getFullWidgetHeight(height: 60) : getWidgetHeight(height: 60),
+                    height: isSplitScreen
+                        ? getFullWidgetHeight(height: 60)
+                        : getWidgetHeight(height: 60),
                     width: kWidth,
                     decoration: BoxDecoration(
                       color: Color(0xFF34445F),
@@ -81,39 +85,73 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                         IconButton(
                             icon: ImageIcon(
                               AssetImage(AllAssets.bottomHome),
-                              color: context.read<AuthState>().currentIndex == 0 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170),
+                              color: context.read<AuthState>().currentIndex == 0
+                                  ? Color(0xFFAAAAAA)
+                                  : Color.fromARGB(132, 170, 170, 170),
                             ),
                             onPressed: () {
                               context.read<AuthState>().changeIndex(0);
-                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                              Navigator.pushReplacement(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          BottomNavigation()));
                             }),
                         IconButton(
                             icon: ImageIcon(AssetImage(AllAssets.bottomPL),
-                                color: context.read<AuthState>().currentIndex == 1 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
+                                color:
+                                    context.read<AuthState>().currentIndex == 1
+                                        ? Color(0xFFAAAAAA)
+                                        : Color.fromARGB(132, 170, 170, 170)),
                             onPressed: () {
                               context.read<AuthState>().changeIndex(1);
-                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                              Navigator.pushReplacement(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          BottomNavigation()));
                             }),
                         IconButton(
                             icon: ImageIcon(AssetImage(AllAssets.bottomIS),
-                                color: context.read<AuthState>().currentIndex == 2 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
+                                color:
+                                    context.read<AuthState>().currentIndex == 2
+                                        ? Color(0xFFAAAAAA)
+                                        : Color.fromARGB(132, 170, 170, 170)),
                             onPressed: () {
                               context.read<AuthState>().changeIndex(2);
-                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                              Navigator.pushReplacement(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          BottomNavigation()));
                             }),
                         IconButton(
                             icon: ImageIcon(AssetImage(AllAssets.bottomPE),
-                                color: context.read<AuthState>().currentIndex == 3 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
+                                color:
+                                    context.read<AuthState>().currentIndex == 3
+                                        ? Color(0xFFAAAAAA)
+                                        : Color.fromARGB(132, 170, 170, 170)),
                             onPressed: () {
                               context.read<AuthState>().changeIndex(3);
-                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                              Navigator.pushReplacement(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          BottomNavigation()));
                             }),
                         IconButton(
                             icon: ImageIcon(AssetImage(AllAssets.bottomPT),
-                                color: context.read<AuthState>().currentIndex == 4 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
+                                color:
+                                    context.read<AuthState>().currentIndex == 4
+                                        ? Color(0xFFAAAAAA)
+                                        : Color.fromARGB(132, 170, 170, 170)),
                             onPressed: () {
                               context.read<AuthState>().changeIndex(4);
-                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                              Navigator.pushReplacement(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          BottomNavigation()));
                             }),
                       ],
                     ),
@@ -125,14 +163,19 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                   Expanded(
                     child: ListView.builder(
                         padding: EdgeInsets.only(
-                            top: isSplitScreen ? getFullWidgetHeight(height: 10) : getWidgetHeight(height: 10),
-                            bottom: isSplitScreen ? getFullWidgetHeight(height: 10) : getWidgetHeight(height: 10)),
+                            top: isSplitScreen
+                                ? getFullWidgetHeight(height: 10)
+                                : getWidgetHeight(height: 10),
+                            bottom: isSplitScreen
+                                ? getFullWidgetHeight(height: 10)
+                                : getWidgetHeight(height: 10)),
                         shrinkWrap: true,
                         controller: controller,
                         itemCount: widget.linkCats.length,
                         // scrollDirection: Axis.horizontal,
                         itemBuilder: (BuildContext context, int index) {
-                          isPlaying = List.generate(widget.linkCats.length, (index) => false.obs);
+                          isPlaying = List.generate(
+                              widget.linkCats.length, (index) => false.obs);
                           print(widget.linkCats[index].toMap());
                           return AutoScrollTag(
                             key: ValueKey(widget.linkCats[index].name),
@@ -155,7 +198,8 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                                 // expKey: expansionTile,
                                 onExpansionChanged: (val) {
                                   if (val) {
-                                    _selectedWordOnClick = widget.linkCats[index].name;
+                                    _selectedWordOnClick =
+                                        widget.linkCats[index].name;
                                     setState(() {});
                                   }
                                 },
@@ -167,7 +211,10 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                                 //   //     _selectedWordOnClick != null &&
                                 //   //         _selectedWordOnClick == _words[index].text);
                                 // },
-                                initiallyExpanded: _selectedWordOnClick != null && _selectedWordOnClick == widget.linkCats[index].name,
+                                initiallyExpanded:
+                                    _selectedWordOnClick != null &&
+                                        _selectedWordOnClick ==
+                                            widget.linkCats[index].name,
                                 isWord: false,
                                 isRefresh: (val) {
                                   // if (val) _getWords(isRefresh: true);
@@ -181,36 +228,68 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                                   Container(
                                     color: Color(0xff293750),
                                     alignment: Alignment.centerLeft,
-                                    padding: EdgeInsets.symmetric(horizontal: 30, vertical: 10),
+                                    padding: EdgeInsets.symmetric(
+                                        horizontal: 30, vertical: 10),
                                     child: Row(
-                                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.spaceEvenly,
                                       children: [
                                         SPW(10),
                                         InkWell(
                                           onTap: () async {
-                                            if (widget.linkCats[index].link1 != null && widget.linkCats[index].link1!.isNotEmpty) {
-                                              print("LinkCheck:${widget.linkCats[index].link1}");
-                                              String? links = widget.linkCats[index].link1;
-                                              arCallSimulationsLinks.add(links!);
-                                              FirebaseFirestore firestore = FirebaseFirestore.instance;
-                                              String userId = await SharedPref.getSavedString('userId');
-                                              DocumentReference arCallDocument = firestore.collection('arCallSimulationsReport').doc(userId);
+                                            if (widget.linkCats[index].link1 !=
+                                                    null &&
+                                                widget.linkCats[index].link1!
+                                                    .isNotEmpty) {
+                                              sessionName =
+                                                  widget.linkCats[index].name!;
+                                              activityName = "Simulation 1";
+                                              print(
+                                                  "LinkCheck:${widget.linkCats[index].link1}");
+                                              String? links =
+                                                  widget.linkCats[index].link1;
+                                              arCallSimulationsLinks
+                                                  .add(links!);
+                                              FirebaseFirestore firestore =
+                                                  FirebaseFirestore.instance;
+                                              String userId = await SharedPref
+                                                  .getSavedString('userId');
+                                              DocumentReference arCallDocument =
+                                                  firestore
+                                                      .collection(
+                                                          'arCallSimulationsReport')
+                                                      .doc(userId);
 
                                               await arCallDocument.update({
-                                                'isLink': FieldValue.arrayUnion([widget.linkCats[index].link1]),
+                                                'isLink':
+                                                    FieldValue.arrayUnion([
+                                                  widget.linkCats[index].link1
+                                                ]),
                                               }).then((_) {
-                                                print('Link added to Firestore: ${widget.linkCats[index].link1}');
+                                                print(
+                                                    'Link added to Firestore: ${widget.linkCats[index].link1}');
                                               }).catchError((e) {
-                                                print('Error updating Firestore: $e');
+                                                print(
+                                                    'Error updating Firestore: $e');
                                               });
-                                              SharedPreferences prefs = await SharedPreferences.getInstance();
-                                              await prefs.setStringList('InAppWebViewPage', [widget.linkCats[index].link1!]);
-                                              await prefs.setString('lastAccess', 'InAppWebViewPage');
+                                              SharedPreferences prefs =
+                                                  await SharedPreferences
+                                                      .getInstance();
+                                              await prefs.setStringList(
+                                                  'InAppWebViewPage', [
+                                                widget.linkCats[index].link1!
+                                              ]);
+                                              await prefs.setString(
+                                                  'lastAccess',
+                                                  'InAppWebViewPage');
                                               Navigator.push(
                                                   context,
                                                   MaterialPageRoute(
-                                                      builder: (context) => InAppWebViewPage(
-                                                            url: widget.linkCats[index].link1!,
+                                                      builder: (context) =>
+                                                          InAppWebViewPage(
+                                                            url: widget
+                                                                .linkCats[index]
+                                                                .link1!,
                                                           )));
                                             }
                                           },
@@ -225,7 +304,11 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                                               SPW(5),
                                               Image.asset(
                                                 AllAssets.interb,
-                                                color: (widget.linkCats[index].link1 != null && widget.linkCats[index].link1!.isNotEmpty)
+                                                color: (widget.linkCats[index]
+                                                                .link1 !=
+                                                            null &&
+                                                        widget.linkCats[index]
+                                                            .link1!.isNotEmpty)
                                                     ? Colors.white
                                                     : Colors.grey,
                                                 width: 25,
@@ -237,29 +320,61 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                                         SPW(15),
                                         InkWell(
                                             onTap: () async {
-                                              if (widget.linkCats[index].link2 != null && widget.linkCats[index].link2!.isNotEmpty) {
-                                                print("LinkCheckkk:${widget.linkCats[index].link2}");
-                                                String? links2 = widget.linkCats[index].link2;
-                                                arCallSimulationsLinks.add(links2!);
-                                                FirebaseFirestore firestore = FirebaseFirestore.instance;
-                                                String userId = await SharedPref.getSavedString('userId');
-                                                DocumentReference softSkills = firestore.collection('arCallSimulationsReport').doc(userId);
+                                              if (widget.linkCats[index]
+                                                          .link2 !=
+                                                      null &&
+                                                  widget.linkCats[index].link2!
+                                                      .isNotEmpty) {
+                                                sessionName = widget
+                                                    .linkCats[index].name!;
+                                                activityName = "Simulation 2";
+                                                print(
+                                                    "LinkCheckkk:${widget.linkCats[index].link2}");
+                                                String? links2 = widget
+                                                    .linkCats[index].link2;
+                                                arCallSimulationsLinks
+                                                    .add(links2!);
+                                                FirebaseFirestore firestore =
+                                                    FirebaseFirestore.instance;
+                                                String userId = await SharedPref
+                                                    .getSavedString('userId');
+                                                DocumentReference softSkills =
+                                                    firestore
+                                                        .collection(
+                                                            'arCallSimulationsReport')
+                                                        .doc(userId);
 
                                                 await softSkills.update({
-                                                  'isLink': FieldValue.arrayUnion([widget.linkCats[index].link2]),
+                                                  'isLink':
+                                                      FieldValue.arrayUnion([
+                                                    widget.linkCats[index].link2
+                                                  ]),
                                                 }).then((_) {
-                                                  print('Link added to Firestore: ${widget.linkCats[index].link2}');
+                                                  print(
+                                                      'Link added to Firestore: ${widget.linkCats[index].link2}');
                                                 }).catchError((e) {
-                                                  print('Error updating Firestore: $e');
+                                                  print(
+                                                      'Error updating Firestore: $e');
                                                 });
-                                                SharedPreferences prefs = await SharedPreferences.getInstance();
-                                                await prefs.setStringList('InAppWebViewPage', [widget.linkCats[index].link2!]);
-                                                await prefs.setString('lastAccess', 'InAppWebViewPage');
+                                                SharedPreferences prefs =
+                                                    await SharedPreferences
+                                                        .getInstance();
+                                                await prefs.setStringList(
+                                                    'InAppWebViewPage', [
+                                                  widget.linkCats[index].link2!
+                                                ]);
+                                                await prefs.setString(
+                                                    'lastAccess',
+                                                    'InAppWebViewPage');
                                                 Navigator.push(
                                                     context,
                                                     MaterialPageRoute(
-                                                        builder: (context) => InAppWebViewPage(
-                                                              url: widget.linkCats[index].link2!,
+                                                        builder: (context) =>
+                                                            InAppWebViewPage(
+                                                              url: widget
+                                                                  .linkCats[
+                                                                      index]
+                                                                  .link2!,
                                                             )));
                                               }
                                             },
@@ -274,7 +389,13 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                                                 SPW(5),
                                                 Image.asset(
                                                   AllAssets.interb,
-                                                  color: (widget.linkCats[index].link2 != null && widget.linkCats[index].link2!.isNotEmpty)
+                                                  color: (widget.linkCats[index]
+                                                                  .link2 !=
+                                                              null &&
+                                                          widget
+                                                              .linkCats[index]
+                                                              .link2!
+                                                              .isNotEmpty)
                                                       ? Colors.white
                                                       : Colors.grey,
                                                   width: 25,
@@ -285,28 +406,60 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                                         SPW(15),
                                         InkWell(
                                             onTap: () async {
-                                              if (widget.linkCats[index].link3 != null && widget.linkCats[index].link3!.isNotEmpty) {
-                                                print("LinkCheckkk:${widget.linkCats[index].link3}");
-                                                String? links3 = widget.linkCats[index].link3;
-                                                arCallSimulationsLinks.add(links3!);
-                                                FirebaseFirestore firestore = FirebaseFirestore.instance;
-                                                String userId = await SharedPref.getSavedString('userId');
-                                                DocumentReference softSkills = firestore.collection('arCallSimulationsReport').doc(userId);
+                                              if (widget.linkCats[index]
+                                                          .link3 !=
+                                                      null &&
+                                                  widget.linkCats[index].link3!
+                                                      .isNotEmpty) {
+                                                sessionName = widget
+                                                    .linkCats[index].name!;
+                                                activityName = "Simulation 3";
+                                                print(
+                                                    "LinkCheckkk:${widget.linkCats[index].link3}");
+                                                String? links3 = widget
+                                                    .linkCats[index].link3;
+                                                arCallSimulationsLinks
+                                                    .add(links3!);
+                                                FirebaseFirestore firestore =
+                                                    FirebaseFirestore.instance;
+                                                String userId = await SharedPref
+                                                    .getSavedString('userId');
+                                                DocumentReference softSkills =
+                                                    firestore
+                                                        .collection(
+                                                            'arCallSimulationsReport')
+                                                        .doc(userId);
                                                 await softSkills.update({
-                                                  'isLink': FieldValue.arrayUnion([widget.linkCats[index].link3]),
+                                                  'isLink':
+                                                      FieldValue.arrayUnion([
+                                                    widget.linkCats[index].link3
+                                                  ]),
                                                 }).then((_) {
-                                                  print('Link added to Firestore: ${widget.linkCats[index].link3}');
+                                                  print(
+                                                      'Link added to Firestore: ${widget.linkCats[index].link3}');
                                                 }).catchError((e) {
-                                                  print('Error updating Firestore: $e');
+                                                  print(
+                                                      'Error updating Firestore: $e');
                                                 });
-                                                SharedPreferences prefs = await SharedPreferences.getInstance();
-                                                await prefs.setStringList('InAppWebViewPage', [widget.linkCats[index].link3!]);
-                                                await prefs.setString('lastAccess', 'InAppWebViewPage');
+                                                SharedPreferences prefs =
+                                                    await SharedPreferences
+                                                        .getInstance();
+                                                await prefs.setStringList(
+                                                    'InAppWebViewPage', [
+                                                  widget.linkCats[index].link3!
+                                                ]);
+                                                await prefs.setString(
+                                                    'lastAccess',
+                                                    'InAppWebViewPage');
                                                 Navigator.push(
                                                     context,
                                                     MaterialPageRoute(
-                                                        builder: (context) => InAppWebViewPage(
-                                                              url: widget.linkCats[index].link3!,
+                                                        builder: (context) =>
+                                                            InAppWebViewPage(
+                                                              url: widget
+                                                                  .linkCats[
+                                                                      index]
+                                                                  .link3!,
                                                             )));
                                               }
                                             },
@@ -321,7 +474,13 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                                                 SPW(5),
                                                 Image.asset(
                                                   AllAssets.interb,
-                                                  color: (widget.linkCats[index].link3 != null && widget.linkCats[index].link3!.isNotEmpty)
+                                                  color: (widget.linkCats[index]
+                                                                  .link3 !=
+                                                              null &&
+                                                          widget
+                                                              .linkCats[index]
+                                                              .link3!
+                                                              .isNotEmpty)
                                                       ? Colors.white
                                                       : Colors.grey,
                                                   width: 25,
@@ -339,7 +498,9 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                         }),
                   ),
                   Container(
-                    height: isSplitScreen ? getFullWidgetHeight(height: 60) : getWidgetHeight(height: 60),
+                    height: isSplitScreen
+                        ? getFullWidgetHeight(height: 60)
+                        : getWidgetHeight(height: 60),
                     width: kWidth,
                     decoration: BoxDecoration(
                       color: Color(0xFF34445F),
@@ -350,39 +511,73 @@ class _ProcessCatScreenState extends State<InteracticeCatScreen> {
                         IconButton(
                             icon: ImageIcon(
                               AssetImage(AllAssets.bottomHome),
-                              color: context.read<AuthState>().currentIndex == 0 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170),
+                              color: context.read<AuthState>().currentIndex == 0
+                                  ? Color(0xFFAAAAAA)
+                                  : Color.fromARGB(132, 170, 170, 170),
                             ),
                             onPressed: () {
                               context.read<AuthState>().changeIndex(0);
-                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                              Navigator.pushReplacement(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          BottomNavigation()));
                             }),
                         IconButton(
                             icon: ImageIcon(AssetImage(AllAssets.bottomPL),
-                                color: context.read<AuthState>().currentIndex == 1 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
+                                color:
+                                    context.read<AuthState>().currentIndex == 1
+                                        ? Color(0xFFAAAAAA)
+                                        : Color.fromARGB(132, 170, 170, 170)),
                             onPressed: () {
                               context.read<AuthState>().changeIndex(1);
-                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                              Navigator.pushReplacement(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          BottomNavigation()));
                             }),
                         IconButton(
                             icon: ImageIcon(AssetImage(AllAssets.bottomIS),
-                                color: context.read<AuthState>().currentIndex == 2 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
+                                color:
+                                    context.read<AuthState>().currentIndex == 2
+                                        ? Color(0xFFAAAAAA)
+                                        : Color.fromARGB(132, 170, 170, 170)),
                             onPressed: () {
                               context.read<AuthState>().changeIndex(2);
-                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                              Navigator.pushReplacement(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          BottomNavigation()));
                             }),
                         IconButton(
                             icon: ImageIcon(AssetImage(AllAssets.bottomPE),
-                                color: context.read<AuthState>().currentIndex == 3 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
+                                color:
+                                    context.read<AuthState>().currentIndex == 3
+                                        ? Color(0xFFAAAAAA)
+                                        : Color.fromARGB(132, 170, 170, 170)),
                             onPressed: () {
                               context.read<AuthState>().changeIndex(3);
-                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                              Navigator.pushReplacement(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          BottomNavigation()));
                             }),
                         IconButton(
                             icon: ImageIcon(AssetImage(AllAssets.bottomPT),
-                                color: context.read<AuthState>().currentIndex == 4 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
+                                color:
+                                    context.read<AuthState>().currentIndex == 4
+                                        ? Color(0xFFAAAAAA)
+                                        : Color.fromARGB(132, 170, 170, 170)),
                             onPressed: () {
                               context.read<AuthState>().changeIndex(4);
-                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                              Navigator.pushReplacement(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          BottomNavigation()));
                             }),
                       ],
                     ),
