@@ -1,13 +1,11 @@
 import 'dart:developer';
 import 'dart:io';
-// import 'package:cloud_firestore/cloud_firestore.dart';
-// import 'package:connectivity_plus/connectivity_plus.dart';
-import 'package:flutter/cupertino.dart';
+
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 // import 'package:flutter/services.dart';
 // import 'package:flutter/widgets.dart';
 import 'package:get/get.dart';
+import 'package:http/http.dart' as http;
 import 'package:intl/intl.dart';
 import 'package:litelearninglab/API/api.dart';
 import 'package:litelearninglab/common_widgets/boom_menu.dart';
@@ -22,11 +20,10 @@ import 'package:litelearninglab/models/ProLab.dart';
 import 'package:litelearninglab/models/Word.dart';
 import 'package:litelearninglab/screens/dialogs/own_word_dialog.dart';
 import 'package:litelearninglab/screens/dialogs/speech_analytics_dialog.dart';
-import 'package:litelearninglab/screens/profluent_english/profluent_english_screen.dart';
-import 'package:litelearninglab/screens/profluent_english/word_screen.dart';
 import 'package:litelearninglab/screens/word_screen/widgets/drop_down_word_item.dart';
 import 'package:litelearninglab/screens/word_screen/widgets/word_menu.dart';
 import 'package:litelearninglab/utils/bottom_navigation.dart';
+import 'package:litelearninglab/utils/commonfunctions/common_functions.dart';
 import 'package:litelearninglab/utils/encrypt_data.dart';
 import 'package:litelearninglab/utils/firebase_helper_RTD.dart';
 import 'package:litelearninglab/utils/shared_pref.dart';
@@ -42,7 +39,6 @@ import '../../states/auth_state.dart';
 import '../../utils/audio_player_manager.dart';
 import '../../utils/firebase_helper.dart';
 import '../profluent_english/new_profluent_english_screen.dart';
-import 'package:http/http.dart' as http;
 
 enum PlayingRouteState { speakers, earpiece }
 
@@ -82,7 +78,8 @@ class WordScreen extends StatefulWidget {
   }
 }
 
-class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, WidgetsBindingObserver {
+class _WordScreenState extends State<WordScreen>
+    with TickerProviderStateMixin, WidgetsBindingObserver {
   final db = new FirebaseHelperRTD();
   FirebaseHelper _firestore = new FirebaseHelper();
   bool openPlay1StopDialog = false;
@@ -123,9 +120,11 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
     print("url : $url");
     try {
       print("responseeeeeeee");
-      var response = await http.post(Uri.parse(url), body: {"userid": userId, "type": action, "word": word});
+      var response = await http.post(Uri.parse(url),
+          body: {"userid": userId, "type": action, "word": word});
 
-      print("response for pronunciation lab report for correct words : ${response.body}");
+      print(
+          "response for pronunciation lab report for correct words : ${response.body}");
     } catch (e) {
       print("error login : $e");
     }
@@ -142,8 +141,13 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
     /* print("scoreeeeetypeee:${widget.score.runtimeType}");
     print("scoreeee:${widget.score}");*/
     try {
-      var response = await http
-          .post(Uri.parse(url), body: {"userid": userId, "practicetype": practiceType, "score": "", "action": "practice", "successCount": successCount});
+      var response = await http.post(Uri.parse(url), body: {
+        "userid": userId,
+        "practicetype": practiceType,
+        "score": "",
+        "action": "practice",
+        "successCount": successCount
+      });
 
       print("response end practice : ${response.body}");
     } catch (e) {
@@ -154,8 +158,13 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
   @override
   void initState() {
     super.initState();
+    // startTimerMainCategory("name");
+    // subCategoryTitile = widget.title;
     WidgetsBinding.instance.addObserver(this);
-    controller = AutoScrollController(viewportBoundaryGetter: () => Rect.fromLTRB(0, 0, 0, MediaQuery.of(context).padding.bottom), axis: Axis.vertical);
+    controller = AutoScrollController(
+        viewportBoundaryGetter: () =>
+            Rect.fromLTRB(0, 0, 0, MediaQuery.of(context).padding.bottom),
+        axis: Axis.vertical);
 
     _getWords(isRefresh: false);
   }
@@ -216,7 +225,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
           print("isPriorityStatus:${_words[i].isPriority}");
           AuthState userDatas = Provider.of<AuthState>(context, listen: false);
 
-          localPath1 = await Utils.downloadFile(userDatas, _words[i].file!, '${_words[i].text}.mp3', '$appDocPath/${widget.load}');
+          localPath1 = await Utils.downloadFile(userDatas, _words[i].file!,
+              '${_words[i].text}.mp3', '$appDocPath/${widget.load}');
           eLocalPath1 = EncryptData.encryptFile(localPath1, userDatas);
           print("localllpathh11:$localPath1");
           print('eLocalPathhh111 :${eLocalPath1}');
@@ -263,7 +273,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
     SharedPreferences prefs = await SharedPreferences.getInstance();
     String userId = await SharedPref.getSavedString('userId');
     bool isFirstTime = await _isFirstTimeUser(userId);
-    bool wordsFirstTime1 = prefs.getBool(widget.title.removeAllWhitespace) ?? true;
+    bool wordsFirstTime1 =
+        prefs.getBool(widget.title.removeAllWhitespace) ?? true;
     print("world ${wordsFirstTime1}");
     setState(() {});
     if (wordsFirstTime1 == true) {
@@ -272,7 +283,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
       setState(() {});
       await addInitialFav();
       await prefs.setBool(widget.title.removeAllWhitespace, false);
-      wordsFirstTime = await prefs.getBool(widget.title.removeAllWhitespace) ?? false;
+      wordsFirstTime =
+          await prefs.getBool(widget.title.removeAllWhitespace) ?? false;
       setState(() {});
     } else {
       print("wdfdjgi:${wordsFirstTime}");
@@ -362,7 +374,10 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
     print("getwordsscalleddddddd>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>");
     print("searchTerm: ${searchTerm}");
     _isLoading = true;
-    if ((searchTerm == null || searchTerm.length == 0) && !isRefresh) setState(() {});
+    startTimerMainCategory("name");
+    sessionName = widget.title;
+    if ((searchTerm == null || searchTerm.length == 0) && !isRefresh)
+      setState(() {});
     _words = [];
     if (widget.load.length > 0) {
       print("show the data from the firebase");
@@ -382,7 +397,10 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
       isPlaying = List.generate(_words.length, (index) => false.obs);
       // isPlaying = List.generate(_words.length, (index) => false);
       for (Word wr in words) {
-        if ((wr.isFav == 1 && widget.filterLoad == null) || (wr.isFav == 1 && widget.filterLoad != null && widget.filterLoad == wr.cat)) {
+        if ((wr.isFav == 1 && widget.filterLoad == null) ||
+            (wr.isFav == 1 &&
+                widget.filterLoad != null &&
+                widget.filterLoad == wr.cat)) {
           _words.add(wr);
           log("${wr.cat}");
           log("${wr.file}");
@@ -391,13 +409,18 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
     }
 
     if (searchTerm != null && searchTerm.length > 0) {
-      _words = _words.where((element) => element.text!.toLowerCase().contains(searchTerm.toLowerCase())).toList();
+      _words = _words
+          .where((element) =>
+              element.text!.toLowerCase().contains(searchTerm.toLowerCase()))
+          .toList();
     }
     if (widget.word != null) _selectedWordOnClick = widget.word?.id;
     _isLoading = false;
 
     if (widget.word != null) {
-      await controller.scrollToIndex(_words.indexWhere((element) => element.text == widget.word?.id), preferPosition: AutoScrollPosition.begin);
+      await controller.scrollToIndex(
+          _words.indexWhere((element) => element.text == widget.word?.id),
+          preferPosition: AutoScrollPosition.begin);
     }
     if (mounted) setState(() {});
   }
@@ -422,9 +445,12 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
           cursorColor: Colors.white,
           decoration: InputDecoration(
             isDense: true,
-            border: UnderlineInputBorder(borderSide: BorderSide(color: Colors.white)),
-            enabledBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.white)),
-            focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.white)),
+            border: UnderlineInputBorder(
+                borderSide: BorderSide(color: Colors.white)),
+            enabledBorder: UnderlineInputBorder(
+                borderSide: BorderSide(color: Colors.white)),
+            focusedBorder: UnderlineInputBorder(
+                borderSide: BorderSide(color: Colors.white)),
             errorBorder: UnderlineInputBorder(),
             disabledBorder: UnderlineInputBorder(),
             hintText: "Search",
@@ -462,7 +488,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
               Navigator.pop(context);
               return;
             }
-            _getWords(searchTerm: _searchQueryController.text, isRefresh: false);
+            _getWords(
+                searchTerm: _searchQueryController.text, isRefresh: false);
           },
         ),
         SPW(displayWidth(context) / 37.5)
@@ -470,7 +497,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
     }
 
     return <Widget>[
-      if (widget.title.trim() == "Priority List" && !isAllPlaying3 || widget.title.trim() == "All Priority List" && !isAllPlaying3)
+      if (widget.title.trim() == "Priority List" && !isAllPlaying3 ||
+          widget.title.trim() == "All Priority List" && !isAllPlaying3)
         _words.length == 0
             ? SizedBox()
             : IconButton(
@@ -492,7 +520,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
                   });
                   _playAll3Times();
                 }),
-      if (widget.title.trim() == "Priority List" && isAllPlaying3 || widget.title.trim() == "All Priority List" && isAllPlaying3)
+      if (widget.title.trim() == "Priority List" && isAllPlaying3 ||
+          widget.title.trim() == "All Priority List" && isAllPlaying3)
         _words.length == 0
             ? SizedBox()
             : IconButton(
@@ -508,7 +537,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
                   // setState(() {});
                 },
               ),
-      if (widget.title.trim() == "Priority List" && !isAllPlaying || widget.title.trim() == "All Priority List" && !isAllPlaying)
+      if (widget.title.trim() == "Priority List" && !isAllPlaying ||
+          widget.title.trim() == "All Priority List" && !isAllPlaying)
         _words.length == 0
             ? SizedBox()
             : IconButton(
@@ -532,7 +562,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
                   _playAll();
                 },
               ),
-      if (widget.title.trim() == "Priority List" && isAllPlaying || widget.title.trim() == "All Priority List" && isAllPlaying)
+      if (widget.title.trim() == "Priority List" && isAllPlaying ||
+          widget.title.trim() == "All Priority List" && isAllPlaying)
         _words.length == 0
             ? SizedBox()
             : IconButton(
@@ -557,7 +588,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
   }
 
   void _startSearch() {
-    ModalRoute.of(context)?.addLocalHistoryEntry(LocalHistoryEntry(onRemove: _stopSearching));
+    ModalRoute.of(context)
+        ?.addLocalHistoryEntry(LocalHistoryEntry(onRemove: _stopSearching));
 
     setState(() {
       _isSearching = true;
@@ -599,7 +631,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
     Get.dialog(
       Container(
         child: Dialog(
-          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(25.0)),
+          shape:
+              RoundedRectangleBorder(borderRadius: BorderRadius.circular(25.0)),
           child: SpeechAnalyticsDialog(
             true,
             isShowDidNotCatch: notCatch,
@@ -610,7 +643,8 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
         ),
       ),
     ).then((value) {
-      if (value != null && value.isCorrect == "true" || value.isCorrect == "false") {
+      if (value != null && value.isCorrect == "true" ||
+          value.isCorrect == "false") {
         print("checkkkkkkkkkkkkkk");
         _selectedWord = word;
         _isCorrect = value.isCorrect == "true" ? true : false;
@@ -620,10 +654,16 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
         setState(() {
           print('wordddddddd: ${word}');
           print("valueIsCorrect: ${_isCorrect}");
-          _isCorrect ? pronunciationLabReport(actionType: "correct", word: word) : "";
           _isCorrect
-              ? endPractice(practiceType: "Pronunciation Sound Lab Report", successCount: "correct")
-              : endPractice(practiceType: "Pronunciation Sound Lab Report", successCount: "wrong");
+              ? pronunciationLabReport(actionType: "correct", word: word)
+              : "";
+          _isCorrect
+              ? endPractice(
+                  practiceType: "Pronunciation Sound Lab Report",
+                  successCount: "correct")
+              : endPractice(
+                  practiceType: "Pronunciation Sound Lab Report",
+                  successCount: "wrong");
         });
       } else if (value != null && value.isCorrect == "notCatch") {
         print("checkk222");
@@ -660,11 +700,15 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
         break; // Exit the loop if paused
       }
 
-      if (isAllPlaying && _words[i].file != null && _words[i].file!.isNotEmpty) {
+      if (isAllPlaying &&
+          _words[i].file != null &&
+          _words[i].file!.isNotEmpty) {
         _words[i].isPlaying = true;
         isPlaying[i].value = true;
         setState(() {});
-        await _audioPlayerManager.play(_words[i].file!, context: context, localPath: _words[i].localPath, decodedPath: (val) {
+        await _audioPlayerManager.play(_words[i].file!,
+            context: context,
+            localPath: _words[i].localPath, decodedPath: (val) {
           eLocalPath = val;
         });
 
@@ -677,8 +721,11 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
             File(eLocalPath!).delete();
           } catch (e) {}
         }
-
+        String company = await SharedPref.getSavedString("companyId");
+        String batch = await SharedPref.getSavedString("batch");
         _firestore.saveWordListReport(
+          companyId: company,
+          batch: batch,
           isPractice: false,
           company: userDatas.appUser!.company!,
           name: userDatas.appUser!.UserMname,
@@ -751,14 +798,16 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
           if (!isAllPlaying3) {
             _currentIndex3 = i;
             _currentPlayCount3 = j;
-            print("Paused at index: $_currentIndex3, play count: $_currentPlayCount3");
+            print(
+                "Paused at index: $_currentIndex3, play count: $_currentPlayCount3");
             return;
           }
           if (_isPaused3) {
             print("sjaoijdifjiwjr");
             _currentIndex3 = i;
             _currentPlayCount3 = j;
-            print("Paused at index: $_currentIndex3, play count: $_currentPlayCount3");
+            print(
+                "Paused at index: $_currentIndex3, play count: $_currentPlayCount3");
             return;
           }
 
@@ -787,8 +836,11 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
               log("Failed to delete file: $e");
             }
           }
-
+          String company = await SharedPref.getSavedString("companyId");
+          String batch = await SharedPref.getSavedString("batch");
           _firestore.saveWordListReport(
+            companyId: company,
+            batch: batch,
             isPractice: false,
             company: userDatas.appUser!.company!,
             name: userDatas.appUser!.UserMname,
@@ -856,477 +908,708 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
 
   @override
   Widget build(BuildContext context) {
-    return BackgroundWidget(
-        appBar: AppBar(
-          leading: IconButton(
-            onPressed: () {
-              print("back button is callingggggg");
-              if (widget.check == null) {
-                print("check is empty");
-                Navigator.pop(context);
-              } else if (widget.check!) {
-                print("checkingggg:${widget.check}");
-                print("checkvjdiv");
-                print("titlecheckkk:${widget.itemWordList![widget.index!]['title']}");
-                Navigator.pop(context);
-                // Navigator.pop(context);
-                // Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) {
-                //   return WordScreen(
-                //     itemWordList: widget.itemWordList,
-                //     controllerList: widget.controllerList,
-                //     title: widget.itemWordList![widget.index!]['title'],
-                //     load: widget.itemWordList![widget.index!]['load'],
-                //   );
-                // }));
-              } else if (widget.check == null && widget.backButtonCheck == null) {
-                print("djgirhivgh");
-                Navigator.pop(context);
-              } else if (widget.check == false && widget.backButtonCheck!) {
-                print("djighrijv");
-                Navigator.pop(context);
-                // Navigator.pop(context);
-                // Navigator.pushReplacement(
-                //     context,
-                //     MaterialPageRoute(
-                //         builder: (context) => WordScreenProfluentEnglish(
-                //               title: widget.checkTitle!,
-                //               load: widget.checkTitle!,
-                //               soundPractice: widget.soundPractice!,
-                //             )));
-              } else {
-                print("calleddddddfgnirfjgirfjgi");
-                Navigator.pop(context);
-              }
-              /*     Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) {
-                return LabS
-              }));*/
-            },
-            icon: Icon(
-              Icons.arrow_back_ios_new_rounded,
+    return PopScope(
+      onPopInvoked: (didPop) {
+        stopTimerMainCategory();
+      },
+      child: BackgroundWidget(
+          appBar: AppBar(
+            leading: IconButton(
+              onPressed: () {
+                stopTimerMainCategory();
+                print("back button is callingggggg");
+                if (widget.check == null) {
+                  print("check is empty");
+                  Navigator.pop(context);
+                } else if (widget.check!) {
+                  print("checkingggg:${widget.check}");
+                  print("checkvjdiv");
+                  print(
+                      "titlecheckkk:${widget.itemWordList![widget.index!]['title']}");
+                  Navigator.pop(context);
+                  // Navigator.pop(context);
+                  // Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) {
+                  //   return WordScreen(
+                  //     itemWordList: widget.itemWordList,
+                  //     controllerList: widget.controllerList,
+                  //     title: widget.itemWordList![widget.index!]['title'],
+                  //     load: widget.itemWordList![widget.index!]['load'],
+                  //   );
+                  // }));
+                } else if (widget.check == null &&
+                    widget.backButtonCheck == null) {
+                  print("djgirhivgh");
+                  Navigator.pop(context);
+                } else if (widget.check == false && widget.backButtonCheck!) {
+                  print("djighrijv");
+                  Navigator.pop(context);
+                  // Navigator.pop(context);
+                  // Navigator.pushReplacement(
+                  //     context,
+                  //     MaterialPageRoute(
+                  //         builder: (context) => WordScreenProfluentEnglish(
+                  //               title: widget.checkTitle!,
+                  //               load: widget.checkTitle!,
+                  //               soundPractice: widget.soundPractice!,
+                  //             )));
+                } else {
+                  print("calleddddddfgnirfjgirfjgi");
+                  Navigator.pop(context);
+                }
+                /*     Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) {
+                  return LabS
+                }));*/
+              },
+              icon: Icon(
+                Icons.arrow_back_ios_new_rounded,
+              ),
             ),
-          ),
-          flexibleSpace: !_isSearching
-              ? Padding(
-                  padding: EdgeInsets.only(left: getWidgetWidth(width: 50)),
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      SizedBox(
-                        width: displayWidth(context) / 1.4,
-                        child: Text(
-                          widget.title,
-                          maxLines: 1,
-                          style: TextStyle(
-                              fontFamily: Keys.fontFamily,
-                              fontSize: globalFontSize(18, context),
-                              color: Colors.white,
-                              fontWeight: FontWeight.w600,
-                              overflow: TextOverflow.ellipsis),
-                        ),
-                      ),
-                    ],
-                  ),
-                )
-              : _buildSearchField(),
-          iconTheme: IconThemeData(color: Colors.white),
-          centerTitle: false,
-          actions: _buildActions(),
-          backgroundColor: Color(0xFF324265),
-        ),
-        body:
-            //  _isConnected == false
-            //     ? Center(
-            //         child: Text(
-            //           "No Network Connection",
-            //           style: TextStyle(
-            //               color: AppColors.white, fontFamily: Keys.fontFamily),
-            //         ),
-            //       )
-            //     :
-            _words.length == 0 && !_isLoading
-                ? Column(
-                    children: [
-                      Expanded(
-                        child: Center(
+            flexibleSpace: !_isSearching
+                ? Padding(
+                    padding: EdgeInsets.only(left: getWidgetWidth(width: 50)),
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        SizedBox(
+                          width: displayWidth(context) / 1.4,
                           child: Text(
-                            "Not Found",
-                            style: TextStyle(color: AppColors.white, fontFamily: Keys.fontFamily),
+                            widget.title,
+                            maxLines: 1,
+                            style: TextStyle(
+                                fontFamily: Keys.fontFamily,
+                                fontSize: globalFontSize(18, context),
+                                color: Colors.white,
+                                fontWeight: FontWeight.w600,
+                                overflow: TextOverflow.ellipsis),
                           ),
                         ),
-                      ),
-                      Container(
-                        height: isSplitScreen ? getFullWidgetHeight(height: 60) : getWidgetHeight(height: 60),
-                        width: kWidth,
-                        decoration: BoxDecoration(
-                          color: Color(0xFF34445F),
+                      ],
+                    ),
+                  )
+                : _buildSearchField(),
+            iconTheme: IconThemeData(color: Colors.white),
+            centerTitle: false,
+            actions: _buildActions(),
+            backgroundColor: Color(0xFF324265),
+          ),
+          body:
+              //  _isConnected == false
+              //     ? Center(
+              //         child: Text(
+              //           "No Network Connection",
+              //           style: TextStyle(
+              //               color: AppColors.white, fontFamily: Keys.fontFamily),
+              //         ),
+              //       )
+              //     :
+              _words.length == 0 && !_isLoading
+                  ? Column(
+                      children: [
+                        Expanded(
+                          child: Center(
+                            child: Text(
+                              "Not Found",
+                              style: TextStyle(
+                                  color: AppColors.white,
+                                  fontFamily: Keys.fontFamily),
+                            ),
+                          ),
                         ),
-                        child: Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                        Container(
+                          height: isSplitScreen
+                              ? getFullWidgetHeight(height: 60)
+                              : getWidgetHeight(height: 60),
+                          width: kWidth,
+                          decoration: BoxDecoration(
+                            color: Color(0xFF34445F),
+                          ),
+                          child: Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                            children: [
+                              IconButton(
+                                  icon: ImageIcon(
+                                    AssetImage(AllAssets.bottomHome),
+                                    color: context
+                                                .read<AuthState>()
+                                                .currentIndex ==
+                                            0
+                                        ? Color(0xFFAAAAAA)
+                                        : Color.fromARGB(132, 170, 170, 170),
+                                  ),
+                                  onPressed: () {
+                                    context.read<AuthState>().changeIndex(0);
+                                    Navigator.pushReplacement(
+                                        context,
+                                        MaterialPageRoute(
+                                            builder: (context) =>
+                                                BottomNavigation()));
+                                  }),
+                              IconButton(
+                                  icon: ImageIcon(
+                                      AssetImage(AllAssets.bottomPL),
+                                      color: context
+                                                  .read<AuthState>()
+                                                  .currentIndex ==
+                                              1
+                                          ? Color(0xFFAAAAAA)
+                                          : Color.fromARGB(132, 170, 170, 170)),
+                                  onPressed: () {
+                                    context.read<AuthState>().changeIndex(1);
+                                    Navigator.pushReplacement(
+                                        context,
+                                        MaterialPageRoute(
+                                            builder: (context) =>
+                                                BottomNavigation()));
+                                  }),
+                              IconButton(
+                                  icon: ImageIcon(
+                                      AssetImage(AllAssets.bottomIS),
+                                      color: context
+                                                  .read<AuthState>()
+                                                  .currentIndex ==
+                                              2
+                                          ? Color(0xFFAAAAAA)
+                                          : Color.fromARGB(132, 170, 170, 170)),
+                                  onPressed: () {
+                                    context.read<AuthState>().changeIndex(2);
+                                    Navigator.pushReplacement(
+                                        context,
+                                        MaterialPageRoute(
+                                            builder: (context) =>
+                                                BottomNavigation()));
+                                  }),
+                              IconButton(
+                                  icon: ImageIcon(
+                                      AssetImage(AllAssets.bottomPE),
+                                      color: context
+                                                  .read<AuthState>()
+                                                  .currentIndex ==
+                                              3
+                                          ? Color(0xFFAAAAAA)
+                                          : Color.fromARGB(132, 170, 170, 170)),
+                                  onPressed: () {
+                                    context.read<AuthState>().changeIndex(3);
+                                    Navigator.pushReplacement(
+                                        context,
+                                        MaterialPageRoute(
+                                            builder: (context) =>
+                                                BottomNavigation()));
+                                  }),
+                              IconButton(
+                                  icon: ImageIcon(
+                                      AssetImage(AllAssets.bottomPT),
+                                      color: context
+                                                  .read<AuthState>()
+                                                  .currentIndex ==
+                                              4
+                                          ? Color(0xFFAAAAAA)
+                                          : Color.fromARGB(132, 170, 170, 170)),
+                                  onPressed: () {
+                                    context.read<AuthState>().changeIndex(4);
+                                    Navigator.pushReplacement(
+                                        context,
+                                        MaterialPageRoute(
+                                            builder: (context) =>
+                                                BottomNavigation()));
+                                  }),
+                            ],
+                          ),
+                        )
+                      ],
+                    )
+                  : _isLoading
+                      ? !wordsFirstTime
+                          ? Center(
+                              child: Column(
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                children: [
+                                  Text(
+                                    "This may take a couple of minutes \n(only during the first time).",
+                                    style: TextStyle(
+                                        color: Colors.white, fontSize: 15),
+                                    textAlign: TextAlign.center,
+                                  ),
+                                  SizedBox(
+                                    height: 7,
+                                  ),
+                                  Text(
+                                    "Please stay on this page, do \nnot go back or close the app.",
+                                    style: TextStyle(
+                                        color: Colors.red, fontSize: 15),
+                                    textAlign: TextAlign.center,
+                                  ),
+                                  SizedBox(height: 13),
+                                  CircularProgressIndicator(
+                                    color: Colors.white,
+                                  ),
+                                ],
+                              ),
+                            )
+                          : Center(
+                              child: CircularProgressIndicator(
+                                color: Colors.white,
+                              ),
+                            )
+                      : Column(
                           children: [
-                            IconButton(
-                                icon: ImageIcon(
-                                  AssetImage(AllAssets.bottomHome),
-                                  color: context.read<AuthState>().currentIndex == 0 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170),
-                                ),
-                                onPressed: () {
-                                  context.read<AuthState>().changeIndex(0);
-                                  Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
-                                }),
-                            IconButton(
-                                icon: ImageIcon(AssetImage(AllAssets.bottomPL),
-                                    color: context.read<AuthState>().currentIndex == 1 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
-                                onPressed: () {
-                                  context.read<AuthState>().changeIndex(1);
-                                  Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
-                                }),
-                            IconButton(
-                                icon: ImageIcon(AssetImage(AllAssets.bottomIS),
-                                    color: context.read<AuthState>().currentIndex == 2 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
-                                onPressed: () {
-                                  context.read<AuthState>().changeIndex(2);
-                                  Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
-                                }),
-                            IconButton(
-                                icon: ImageIcon(AssetImage(AllAssets.bottomPE),
-                                    color: context.read<AuthState>().currentIndex == 3 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
-                                onPressed: () {
-                                  context.read<AuthState>().changeIndex(3);
-                                  Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
-                                }),
-                            IconButton(
-                                icon: ImageIcon(AssetImage(AllAssets.bottomPT),
-                                    color: context.read<AuthState>().currentIndex == 4 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
-                                onPressed: () {
-                                  context.read<AuthState>().changeIndex(4);
-                                  Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
-                                }),
+                            Expanded(
+                              child: Stack(
+                                children: [
+                                  ListView.builder(
+                                      padding: EdgeInsets.only(
+                                          top: isSplitScreen
+                                              ? getFullWidgetHeight(height: 10)
+                                              : getWidgetHeight(height: 10)),
+                                      itemCount: _words.length,
+                                      controller: controller,
+                                      itemBuilder:
+                                          (BuildContext context, int index) {
+                                        if (!isAllPlaying && !isAllPlaying3) {
+                                          print(
+                                              "listWordLength:${_words.length}");
+                                          print("dfjusfeiujfejdsfj");
+                                          print(
+                                              "currentIndex : ${_currentIndex}");
+                                          print(
+                                              "wordLengthIndex:${_words.length - 1}");
+                                          isPlaying = List.generate(
+                                              _words.length,
+                                              (index) => false.obs);
+
+                                          /* if (_currentIndex == _words.length - 1) {
+                                            print("sucessss>>>>>>>>>>>>>>>>>>");
+                                            closePlay1StopDialog = true;
+                                          }*/
+                                          //_getWords(isRefresh: false);
+                                        }
+
+                                        return AutoScrollTag(
+                                          key: ValueKey(_words[index].text),
+                                          controller: controller,
+                                          index: index,
+                                          child: Padding(
+                                            padding: EdgeInsets.only(
+                                                bottom: index ==
+                                                        _words.length - 1
+                                                    ? isSplitScreen
+                                                        ? getFullWidgetHeight(
+                                                            height: 60)
+                                                        : getWidgetHeight(
+                                                            height: 60)
+                                                    : 0),
+                                            child: DropDownWordItem(
+                                              localPath:
+                                                  _words[index].localPath,
+                                              load: widget.load,
+                                              length: _words.length,
+                                              index: index,
+                                              // isPlaying: _words[index].isPlaying,
+                                              isDownloaded:
+                                                  (_words[index].localPath !=
+                                                          null &&
+                                                      _words[index]
+                                                          .localPath!
+                                                          .isNotEmpty),
+                                              maintitle: widget.title,
+                                              onExpansionChanged: (val) {
+                                                print(
+                                                    "check1111111111111111111111111444>");
+                                                print(
+                                                    "djfdjfihdifhidfid:${_words[index].localPath}");
+                                                setState(() {
+                                                  _selectedWord = '';
+                                                });
+                                                if (val) {
+                                                  _selectedWordOnClick =
+                                                      _words[index].text;
+                                                  print(
+                                                      "sajidjg:${_words[index].text}");
+                                                  setState(() {});
+                                                  if (_words.length - 2 <=
+                                                      index) {
+                                                    WidgetsBinding.instance
+                                                        .addPostFrameCallback(
+                                                            (_) {
+                                                      _scrollToItem(index);
+                                                    });
+                                                  }
+                                                }
+                                              },
+                                              initiallyExpanded:
+                                                  _selectedWordOnClick !=
+                                                          null &&
+                                                      _selectedWordOnClick ==
+                                                          _words[index].text,
+                                              isWord: true,
+                                              isRefresh: (val) {
+                                                if (val)
+                                                  _getWords(isRefresh: true);
+                                              },
+                                              // words: _words,
+                                              wordId: _words[index].id!,
+                                              isFav: _words[index].isFav!,
+                                              title: _words[index].text!,
+                                              url: _words[index].file,
+                                              onTapForThreePlayerStop:
+                                                  updateThreePlayerFlag,
+                                              children: [
+                                                WordMenu(
+                                                  pronun: _words[index].pronun!,
+                                                  selectedWord: _selectedWord,
+                                                  isCorrect: _selectedWord ==
+                                                          _words[index].text &&
+                                                      _isCorrect,
+                                                  text: _words[index].text!,
+                                                  syllables:
+                                                      _words[index].syllables!,
+                                                  onTapHeadphone: () async {},
+                                                  url: _words[index].file,
+                                                  onTapMic: () async {
+                                                    _showDialog(
+                                                        _words[index].text!,
+                                                        false,
+                                                        context);
+                                                  },
+                                                )
+                                              ],
+                                            ),
+                                          ),
+                                        );
+                                      }),
+                                  openPlay3StopDialog && !closePlay3StopDialog
+                                      ? Positioned(
+                                          right: 120,
+                                          top: 5,
+                                          child: Container(
+                                            height: isSplitScreen
+                                                ? getFullWidgetHeight(
+                                                    height: 45)
+                                                : getWidgetHeight(height: 45),
+                                            width: getWidgetWidth(width: 108),
+                                            decoration: BoxDecoration(
+                                                color: Colors.white,
+                                                borderRadius:
+                                                    BorderRadius.circular(20)),
+                                            child: Row(
+                                              mainAxisAlignment:
+                                                  MainAxisAlignment.spaceEvenly,
+                                              children: [
+                                                Padding(
+                                                  padding:
+                                                      const EdgeInsets.only(
+                                                          left: 8, top: 5),
+                                                  child: Image.asset(
+                                                    "assets/images/Play_thrice_icon.png",
+                                                    // height: 30,
+                                                    width: 30,
+                                                  ),
+                                                ),
+                                                !isAllPlaying3
+                                                    ? audioLoading
+                                                        ? Padding(
+                                                            padding:
+                                                                const EdgeInsets
+                                                                    .all(8.0),
+                                                            child: SizedBox(
+                                                              height: 20,
+                                                              width: 20,
+                                                              child:
+                                                                  CircularProgressIndicator(
+                                                                color: Color(
+                                                                    0XFF34425D),
+                                                                strokeWidth: 2,
+                                                              ),
+                                                            ),
+                                                          )
+                                                        : InkWell(
+                                                            onTap: () {
+                                                              resumeAll3();
+                                                            },
+                                                            child: Icon(Icons.play_arrow,
+                                                                color: Color(
+                                                                    0XFF34425D),
+                                                                size: 30))
+                                                    : audioLoading
+                                                        ? Padding(
+                                                            padding:
+                                                                const EdgeInsets
+                                                                    .all(8.0),
+                                                            child: SizedBox(
+                                                              height: 20,
+                                                              width: 20,
+                                                              child:
+                                                                  CircularProgressIndicator(
+                                                                color: Color(
+                                                                    0XFF34425D),
+                                                                strokeWidth: 2,
+                                                              ),
+                                                            ),
+                                                          )
+                                                        : InkWell(
+                                                            onTap: () {
+                                                              pauseAll3();
+                                                            },
+                                                            child: Icon(
+                                                                Icons.pause,
+                                                                color: Color(
+                                                                    0XFF34425D),
+                                                                size: 30)),
+                                                IconButton(
+                                                    onPressed: () {
+                                                      _audioPlayerManager
+                                                          .stop();
+                                                      isAllPlaying3 = false;
+                                                      setState(() {
+                                                        openPlay3StopDialog =
+                                                            false;
+                                                      });
+                                                    },
+                                                    icon: Icon(
+                                                      Icons.stop,
+                                                      color: Color(0XFF34425D),
+                                                      size: 26,
+                                                    )),
+                                              ],
+                                            ),
+                                          ))
+                                      : openPlay1StopDialog &&
+                                              !closePlay1StopDialog
+                                          ? Positioned(
+                                              right: 120,
+                                              top: 5,
+                                              child: Container(
+                                                height: isSplitScreen
+                                                    ? getFullWidgetHeight(
+                                                        height: 45)
+                                                    : getWidgetHeight(
+                                                        height: 45),
+                                                width:
+                                                    getWidgetWidth(width: 108),
+                                                decoration: BoxDecoration(
+                                                    color: Colors.white,
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            20)),
+                                                child: Row(
+                                                  mainAxisAlignment:
+                                                      MainAxisAlignment
+                                                          .spaceEvenly,
+                                                  children: [
+                                                    Padding(
+                                                      padding:
+                                                          const EdgeInsets.only(
+                                                              left: 8, top: 5),
+                                                      child: Image.asset(
+                                                        "assets/images/Play_once_icon.png",
+                                                        // height: 15,
+                                                        width: 30,
+                                                      ),
+                                                    ),
+                                                    _isPaused
+                                                        ? audioLoading
+                                                            ? Padding(
+                                                                padding:
+                                                                    const EdgeInsets
+                                                                        .all(
+                                                                        8.0),
+                                                                child: SizedBox(
+                                                                  height: 20,
+                                                                  width: 20,
+                                                                  child:
+                                                                      CircularProgressIndicator(
+                                                                    color: Color(
+                                                                        0XFF34425D),
+                                                                    strokeWidth:
+                                                                        2,
+                                                                  ),
+                                                                ),
+                                                              )
+                                                            : InkWell(
+                                                                onTap: () {
+                                                                  print(
+                                                                      "checkkkkkkkkk11");
+                                                                  resumeAll();
+                                                                },
+                                                                child: Icon(
+                                                                    Icons
+                                                                        .play_arrow,
+                                                                    color: Color(
+                                                                        0XFF34425D),
+                                                                    size: 30))
+                                                        : audioLoading
+                                                            ? Padding(
+                                                                padding:
+                                                                    const EdgeInsets
+                                                                        .all(
+                                                                        8.0),
+                                                                child: SizedBox(
+                                                                  height: 20,
+                                                                  width: 20,
+                                                                  child:
+                                                                      CircularProgressIndicator(
+                                                                    color: Color(
+                                                                        0XFF34425D),
+                                                                    strokeWidth:
+                                                                        2,
+                                                                  ),
+                                                                ),
+                                                              )
+                                                            : InkWell(
+                                                                onTap: () {
+                                                                  print(
+                                                                      "checkkk222");
+                                                                  pauseAll();
+                                                                },
+                                                                child: Icon(
+                                                                  Icons.pause,
+                                                                  color: Color(
+                                                                      0XFF34425D),
+                                                                  size: 30,
+                                                                )),
+                                                    InkWell(
+                                                        onTap: () {
+                                                          _audioPlayerManager
+                                                              .stop();
+                                                          isAllPlaying = false;
+                                                          setState(() {
+                                                            openPlay1StopDialog =
+                                                                false;
+                                                          });
+                                                        },
+                                                        child: Icon(
+                                                          Icons.stop,
+                                                          color:
+                                                              Color(0XFF34425D),
+                                                          size: 30,
+                                                        )),
+                                                  ],
+                                                ),
+                                              ))
+                                          : SizedBox()
+                                ],
+                              ),
+                            ),
+                            Container(
+                              height: isSplitScreen
+                                  ? getFullWidgetHeight(height: 60)
+                                  : getWidgetHeight(height: 60),
+                              width: kWidth,
+                              decoration: BoxDecoration(
+                                color: Color(0xFF34445F),
+                              ),
+                              child: Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceEvenly,
+                                children: [
+                                  IconButton(
+                                      icon: ImageIcon(
+                                        AssetImage(AllAssets.bottomHome),
+                                        color: context
+                                                    .read<AuthState>()
+                                                    .currentIndex ==
+                                                0
+                                            ? Color(0xFFAAAAAA)
+                                            : Color.fromARGB(
+                                                132, 170, 170, 170),
+                                      ),
+                                      onPressed: () {
+                                        context
+                                            .read<AuthState>()
+                                            .changeIndex(0);
+                                        Navigator.pushReplacement(
+                                            context,
+                                            MaterialPageRoute(
+                                                builder: (context) =>
+                                                    BottomNavigation()));
+                                      }),
+                                  IconButton(
+                                      icon: ImageIcon(
+                                          AssetImage(AllAssets.bottomPL),
+                                          color: context
+                                                      .read<AuthState>()
+                                                      .currentIndex ==
+                                                  1
+                                              ? Color(0xFFAAAAAA)
+                                              : Color.fromARGB(
+                                                  132, 170, 170, 170)),
+                                      onPressed: () {
+                                        context
+                                            .read<AuthState>()
+                                            .changeIndex(1);
+                                        Navigator.pushReplacement(
+                                            context,
+                                            MaterialPageRoute(
+                                                builder: (context) =>
+                                                    BottomNavigation()));
+                                      }),
+                                  IconButton(
+                                      icon: ImageIcon(
+                                          AssetImage(AllAssets.bottomIS),
+                                          color: context
+                                                      .read<AuthState>()
+                                                      .currentIndex ==
+                                                  2
+                                              ? Color(0xFFAAAAAA)
+                                              : Color.fromARGB(
+                                                  132, 170, 170, 170)),
+                                      onPressed: () {
+                                        context
+                                            .read<AuthState>()
+                                            .changeIndex(2);
+                                        Navigator.pushReplacement(
+                                            context,
+                                            MaterialPageRoute(
+                                                builder: (context) =>
+                                                    BottomNavigation()));
+                                      }),
+                                  IconButton(
+                                      icon: ImageIcon(
+                                          AssetImage(AllAssets.bottomPE),
+                                          color: context
+                                                      .read<AuthState>()
+                                                      .currentIndex ==
+                                                  3
+                                              ? Color(0xFFAAAAAA)
+                                              : Color.fromARGB(
+                                                  132, 170, 170, 170)),
+                                      onPressed: () {
+                                        context
+                                            .read<AuthState>()
+                                            .changeIndex(3);
+                                        Navigator.pushReplacement(
+                                            context,
+                                            MaterialPageRoute(
+                                                builder: (context) =>
+                                                    BottomNavigation()));
+                                      }),
+                                  IconButton(
+                                      icon: ImageIcon(
+                                          AssetImage(AllAssets.bottomPT),
+                                          color: context
+                                                      .read<AuthState>()
+                                                      .currentIndex ==
+                                                  4
+                                              ? Color(0xFFAAAAAA)
+                                              : Color.fromARGB(
+                                                  132, 170, 170, 170)),
+                                      onPressed: () {
+                                        context
+                                            .read<AuthState>()
+                                            .changeIndex(4);
+                                        Navigator.pushReplacement(
+                                            context,
+                                            MaterialPageRoute(
+                                                builder: (context) =>
+                                                    BottomNavigation()));
+                                      }),
+                                ],
+                              ),
+                            )
                           ],
                         ),
-                      )
-                    ],
-                  )
-                : _isLoading
-                    ? !wordsFirstTime
-                        ? Center(
-                            child: Column(
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              children: [
-                                Text(
-                                  "This may take a couple of minutes \n(only during the first time).",
-                                  style: TextStyle(color: Colors.white, fontSize: 15),
-                                  textAlign: TextAlign.center,
-                                ),
-                                SizedBox(
-                                  height: 7,
-                                ),
-                                Text(
-                                  "Please stay on this page, do \nnot go back or close the app.",
-                                  style: TextStyle(color: Colors.red, fontSize: 15),
-                                  textAlign: TextAlign.center,
-                                ),
-                                SizedBox(height: 13),
-                                CircularProgressIndicator(
-                                  color: Colors.white,
-                                ),
-                              ],
-                            ),
-                          )
-                        : Center(
-                            child: CircularProgressIndicator(
-                              color: Colors.white,
-                            ),
-                          )
-                    : Column(
-                        children: [
-                          Expanded(
-                            child: Stack(
-                              children: [
-                                ListView.builder(
-                                    padding: EdgeInsets.only(top: isSplitScreen ? getFullWidgetHeight(height: 10) : getWidgetHeight(height: 10)),
-                                    itemCount: _words.length,
-                                    controller: controller,
-                                    itemBuilder: (BuildContext context, int index) {
-                                      if (!isAllPlaying && !isAllPlaying3) {
-                                        print("listWordLength:${_words.length}");
-                                        print("dfjusfeiujfejdsfj");
-                                        print("currentIndex : ${_currentIndex}");
-                                        print("wordLengthIndex:${_words.length - 1}");
-                                        isPlaying = List.generate(_words.length, (index) => false.obs);
-
-                                        /* if (_currentIndex == _words.length - 1) {
-                                          print("sucessss>>>>>>>>>>>>>>>>>>");
-                                          closePlay1StopDialog = true;
-                                        }*/
-                                        //_getWords(isRefresh: false);
-                                      }
-                                      return AutoScrollTag(
-                                        key: ValueKey(_words[index].text),
-                                        controller: controller,
-                                        index: index,
-                                        child: Padding(
-                                          padding: EdgeInsets.only(
-                                              bottom: index == _words.length - 1
-                                                  ? isSplitScreen
-                                                      ? getFullWidgetHeight(height: 60)
-                                                      : getWidgetHeight(height: 60)
-                                                  : 0),
-                                          child: DropDownWordItem(
-                                            localPath: _words[index].localPath,
-                                            load: widget.load,
-                                            length: _words.length,
-                                            index: index,
-                                            // isPlaying: _words[index].isPlaying,
-                                            isDownloaded: (_words[index].localPath != null && _words[index].localPath!.isNotEmpty),
-                                            maintitle: widget.title,
-                                            onExpansionChanged: (val) {
-                                              print("check1111111111111111111111111444>");
-                                              print("djfdjfihdifhidfid:${_words[index].localPath}");
-                                              setState(() {
-                                                _selectedWord = '';
-                                              });
-                                              if (val) {
-                                                _selectedWordOnClick = _words[index].text;
-                                                print("sajidjg:${_words[index].text}");
-                                                setState(() {});
-                                                if (_words.length - 2 <= index) {
-                                                  WidgetsBinding.instance.addPostFrameCallback((_) {
-                                                    _scrollToItem(index);
-                                                  });
-                                                }
-                                              }
-                                            },
-                                            initiallyExpanded: _selectedWordOnClick != null && _selectedWordOnClick == _words[index].text,
-                                            isWord: true,
-                                            isRefresh: (val) {
-                                              if (val) _getWords(isRefresh: true);
-                                            },
-                                            // words: _words,
-                                            wordId: _words[index].id!,
-                                            isFav: _words[index].isFav!,
-                                            title: _words[index].text!,
-                                            url: _words[index].file,
-                                            onTapForThreePlayerStop: updateThreePlayerFlag,
-                                            children: [
-                                              WordMenu(
-                                                pronun: _words[index].pronun!,
-                                                selectedWord: _selectedWord,
-                                                isCorrect: _selectedWord == _words[index].text && _isCorrect,
-                                                text: _words[index].text!,
-                                                syllables: _words[index].syllables!,
-                                                onTapHeadphone: () async {},
-                                                url: _words[index].file,
-                                                onTapMic: () async {
-                                                  _showDialog(_words[index].text!, false, context);
-                                                },
-                                              )
-                                            ],
-                                          ),
-                                        ),
-                                      );
-                                    }),
-                                openPlay3StopDialog && !closePlay3StopDialog
-                                    ? Positioned(
-                                        right: 120,
-                                        top: 5,
-                                        child: Container(
-                                          height: isSplitScreen ? getFullWidgetHeight(height: 45) : getWidgetHeight(height: 45),
-                                          width: getWidgetWidth(width: 108),
-                                          decoration: BoxDecoration(color: Colors.white, borderRadius: BorderRadius.circular(20)),
-                                          child: Row(
-                                            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                                            children: [
-                                              Padding(
-                                                padding: const EdgeInsets.only(left: 8, top: 5),
-                                                child: Image.asset(
-                                                  "assets/images/Play_thrice_icon.png",
-                                                  // height: 30,
-                                                  width: 30,
-                                                ),
-                                              ),
-                                              !isAllPlaying3
-                                                  ? audioLoading
-                                                      ? Padding(
-                                                          padding: const EdgeInsets.all(8.0),
-                                                          child: SizedBox(
-                                                            height: 20,
-                                                            width: 20,
-                                                            child: CircularProgressIndicator(
-                                                              color: Color(0XFF34425D),
-                                                              strokeWidth: 2,
-                                                            ),
-                                                          ),
-                                                        )
-                                                      : InkWell(
-                                                          onTap: () {
-                                                            resumeAll3();
-                                                          },
-                                                          child: Icon(Icons.play_arrow, color: Color(0XFF34425D), size: 30))
-                                                  : audioLoading
-                                                      ? Padding(
-                                                          padding: const EdgeInsets.all(8.0),
-                                                          child: SizedBox(
-                                                            height: 20,
-                                                            width: 20,
-                                                            child: CircularProgressIndicator(
-                                                              color: Color(0XFF34425D),
-                                                              strokeWidth: 2,
-                                                            ),
-                                                          ),
-                                                        )
-                                                      : InkWell(
-                                                          onTap: () {
-                                                            pauseAll3();
-                                                          },
-                                                          child: Icon(Icons.pause, color: Color(0XFF34425D), size: 30)),
-                                              IconButton(
-                                                  onPressed: () {
-                                                    _audioPlayerManager.stop();
-                                                    isAllPlaying3 = false;
-                                                    setState(() {
-                                                      openPlay3StopDialog = false;
-                                                    });
-                                                  },
-                                                  icon: Icon(
-                                                    Icons.stop,
-                                                    color: Color(0XFF34425D),
-                                                    size: 26,
-                                                  )),
-                                            ],
-                                          ),
-                                        ))
-                                    : openPlay1StopDialog && !closePlay1StopDialog
-                                        ? Positioned(
-                                            right: 120,
-                                            top: 5,
-                                            child: Container(
-                                              height: isSplitScreen ? getFullWidgetHeight(height: 45) : getWidgetHeight(height: 45),
-                                              width: getWidgetWidth(width: 108),
-                                              decoration: BoxDecoration(color: Colors.white, borderRadius: BorderRadius.circular(20)),
-                                              child: Row(
-                                                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                                                children: [
-                                                  Padding(
-                                                    padding: const EdgeInsets.only(left: 8, top: 5),
-                                                    child: Image.asset(
-                                                      "assets/images/Play_once_icon.png",
-                                                      // height: 15,
-                                                      width: 30,
-                                                    ),
-                                                  ),
-                                                  _isPaused
-                                                      ? audioLoading
-                                                          ? Padding(
-                                                              padding: const EdgeInsets.all(8.0),
-                                                              child: SizedBox(
-                                                                height: 20,
-                                                                width: 20,
-                                                                child: CircularProgressIndicator(
-                                                                  color: Color(0XFF34425D),
-                                                                  strokeWidth: 2,
-                                                                ),
-                                                              ),
-                                                            )
-                                                          : InkWell(
-                                                              onTap: () {
-                                                                print("checkkkkkkkkk11");
-                                                                resumeAll();
-                                                              },
-                                                              child: Icon(Icons.play_arrow, color: Color(0XFF34425D), size: 30))
-                                                      : audioLoading
-                                                          ? Padding(
-                                                              padding: const EdgeInsets.all(8.0),
-                                                              child: SizedBox(
-                                                                height: 20,
-                                                                width: 20,
-                                                                child: CircularProgressIndicator(
-                                                                  color: Color(0XFF34425D),
-                                                                  strokeWidth: 2,
-                                                                ),
-                                                              ),
-                                                            )
-                                                          : InkWell(
-                                                              onTap: () {
-                                                                print("checkkk222");
-                                                                pauseAll();
-                                                              },
-                                                              child: Icon(
-                                                                Icons.pause,
-                                                                color: Color(0XFF34425D),
-                                                                size: 30,
-                                                              )),
-                                                  InkWell(
-                                                      onTap: () {
-                                                        _audioPlayerManager.stop();
-                                                        isAllPlaying = false;
-                                                        setState(() {
-                                                          openPlay1StopDialog = false;
-                                                        });
-                                                      },
-                                                      child: Icon(
-                                                        Icons.stop,
-                                                        color: Color(0XFF34425D),
-                                                        size: 30,
-                                                      )),
-                                                ],
-                                              ),
-                                            ))
-                                        : SizedBox()
-                              ],
-                            ),
-                          ),
-                          Container(
-                            height: isSplitScreen ? getFullWidgetHeight(height: 60) : getWidgetHeight(height: 60),
-                            width: kWidth,
-                            decoration: BoxDecoration(
-                              color: Color(0xFF34445F),
-                            ),
-                            child: Row(
-                              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                              children: [
-                                IconButton(
-                                    icon: ImageIcon(
-                                      AssetImage(AllAssets.bottomHome),
-                                      color: context.read<AuthState>().currentIndex == 0 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170),
-                                    ),
-                                    onPressed: () {
-                                      context.read<AuthState>().changeIndex(0);
-                                      Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
-                                    }),
-                                IconButton(
-                                    icon: ImageIcon(AssetImage(AllAssets.bottomPL),
-                                        color: context.read<AuthState>().currentIndex == 1 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
-                                    onPressed: () {
-                                      context.read<AuthState>().changeIndex(1);
-                                      Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
-                                    }),
-                                IconButton(
-                                    icon: ImageIcon(AssetImage(AllAssets.bottomIS),
-                                        color: context.read<AuthState>().currentIndex == 2 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
-                                    onPressed: () {
-                                      context.read<AuthState>().changeIndex(2);
-                                      Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
-                                    }),
-                                IconButton(
-                                    icon: ImageIcon(AssetImage(AllAssets.bottomPE),
-                                        color: context.read<AuthState>().currentIndex == 3 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
-                                    onPressed: () {
-                                      context.read<AuthState>().changeIndex(3);
-                                      Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
-                                    }),
-                                IconButton(
-                                    icon: ImageIcon(AssetImage(AllAssets.bottomPT),
-                                        color: context.read<AuthState>().currentIndex == 4 ? Color(0xFFAAAAAA) : Color.fromARGB(132, 170, 170, 170)),
-                                    onPressed: () {
-                                      context.read<AuthState>().changeIndex(4);
-                                      Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
-                                    }),
-                              ],
-                            ),
-                          )
-                        ],
-                      ),
-        floatingActionButton: _isLoading ? SizedBox() : buildBoomMenu());
+          floatingActionButton: _isLoading ? SizedBox() : buildBoomMenu()),
+    );
   }
 
   buildBoomMenu() {
@@ -1348,12 +1631,18 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
             bm.MenuItem(
               child: Container(
                 padding: EdgeInsets.symmetric(
-                    horizontal: getWidgetWidth(width: 12), vertical: isSplitScreen ? getFullWidgetHeight(height: 12) : getWidgetHeight(height: 12)),
-                decoration: BoxDecoration(color: Colors.white, shape: BoxShape.circle),
+                    horizontal: getWidgetWidth(width: 12),
+                    vertical: isSplitScreen
+                        ? getFullWidgetHeight(height: 12)
+                        : getWidgetHeight(height: 12)),
+                decoration:
+                    BoxDecoration(color: Colors.white, shape: BoxShape.circle),
                 child: Icon(
                   Icons.home,
                   color: Colors.grey,
-                  size: isSplitScreen ? getFullWidgetHeight(height: 18) : getWidgetHeight(height: 18),
+                  size: isSplitScreen
+                      ? getFullWidgetHeight(height: 18)
+                      : getWidgetHeight(height: 18),
                 ),
               ),
               title: "Home",
@@ -1361,18 +1650,27 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
               backgroundColor: Color(0x00000000),
               onTap: () {
                 context.read<AuthState>().changeIndex(0);
-                Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => BottomNavigation()));
+                Navigator.pushReplacement(
+                    context,
+                    MaterialPageRoute(
+                        builder: (context) => BottomNavigation()));
               },
             ),
             bm.MenuItem(
               child: Container(
                 padding: EdgeInsets.symmetric(
-                    horizontal: getWidgetWidth(width: 12), vertical: isSplitScreen ? getFullWidgetHeight(height: 12) : getWidgetHeight(height: 12)),
-                decoration: BoxDecoration(color: Colors.white, shape: BoxShape.circle),
+                    horizontal: getWidgetWidth(width: 12),
+                    vertical: isSplitScreen
+                        ? getFullWidgetHeight(height: 12)
+                        : getWidgetHeight(height: 12)),
+                decoration:
+                    BoxDecoration(color: Colors.white, shape: BoxShape.circle),
                 child: Icon(
                   Icons.keyboard,
                   color: Colors.grey,
-                  size: isSplitScreen ? getFullWidgetHeight(height: 18) : getWidgetHeight(height: 18),
+                  size: isSplitScreen
+                      ? getFullWidgetHeight(height: 18)
+                      : getWidgetHeight(height: 18),
                 ),
               ),
               title: "Try Unlisted Words",
@@ -1398,8 +1696,12 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
             bm.MenuItem(
               child: Container(
                   padding: EdgeInsets.symmetric(
-                      horizontal: getWidgetWidth(width: 12), vertical: isSplitScreen ? getFullWidgetHeight(height: 12) : getWidgetHeight(height: 12)),
-                  decoration: BoxDecoration(color: Colors.white, shape: BoxShape.circle),
+                      horizontal: getWidgetWidth(width: 12),
+                      vertical: isSplitScreen
+                          ? getFullWidgetHeight(height: 12)
+                          : getWidgetHeight(height: 12)),
+                  decoration: BoxDecoration(
+                      color: Colors.white, shape: BoxShape.circle),
                   child: Image.asset("assets/images/filter.png",
                       color: Colors
                           .grey) /*
@@ -1439,8 +1741,10 @@ class _WordScreenState extends State<WordScreen> with TickerProviderStateMixin, 
             bm.MenuItem(
               child: Container(
                   padding: EdgeInsets.all(12),
-                  decoration: BoxDecoration(color: Colors.white, shape: BoxShape.circle),
-                  child: Image.asset("assets/images/filter_all.png", color: Colors.grey)),
+                  decoration: BoxDecoration(
+                      color: Colors.white, shape: BoxShape.circle),
+                  child: Image.asset("assets/images/filter_all.png",
+                      color: Colors.grey)),
               title: "Filter All Priority",
               titleColor: Colors.white,
               backgroundColor: Colors.transparent,
